oops, I didn't see this warning
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@7454 d7cf8633-e32d-0410-b094-e92efae38249

+++ snd_ogg.c
@@ -546,7 +546,7 @@ static const snd_buffer_t* OGG_FetchSound (channel_t* ch, unsigned int* start, u
 		return NULL;
 	}
 	newlength = per_sfx->format.speed * factor;  // -> 1 sec of sound before resampling
-	if(newlength > sizeof(resampling_buffer))
+	if(newlength > (int)sizeof(resampling_buffer))
 		newlength = sizeof(resampling_buffer);
 
 	// Decompress in the resampling_buffer
